clean up style class
