+++ src/engraving/CMakeLists.txt
@@ -83,6 +83,8 @@ set(MODULE_SRC
 
     ${CMAKE_CURRENT_LIST_DIR}/compat/mscxcompat.cpp
     ${CMAKE_CURRENT_LIST_DIR}/compat/mscxcompat.h
+    ${CMAKE_CURRENT_LIST_DIR}/compat/pageformat.cpp
+    ${CMAKE_CURRENT_LIST_DIR}/compat/pageformat.h
 
     ${CMAKE_CURRENT_LIST_DIR}/style/styledef.cpp
     ${CMAKE_CURRENT_LIST_DIR}/style/styledef.h
