+++ src/engraving/libmscore/read206.cpp
@@ -25,6 +25,7 @@
 #include <QRegularExpression>
 
 #include "style/style.h"
+#include "compat/pageformat.h"
 
 #include "xml.h"
 #include "score.h"
@@ -84,6 +85,7 @@
 #include "marker.h"
 
 using namespace mu;
+using namespace mu::engraving;
 
 namespace Ms {
 static void readText206(XmlReader& e, TextBase* t, Element* be);
@@ -96,58 +98,6 @@ static void readText206(XmlReader& e, TextBase* t, Element* be);
 //---------------------------------------------------------
 
 static std::map<QString, std::map<Sid, QVariant> > excessTextStyles206;
-
-//---------------------------------------------------------
-//   setPageFormat
-//    set Style from PageFormat
-//---------------------------------------------------------
-
-void setPageFormat(MStyle* style, const PageFormat& pf)
-{
-    style->set(Sid::pageWidth,            pf.size().width());
-    style->set(Sid::pageHeight,           pf.size().height());
-    style->set(Sid::pagePrintableWidth,   pf.printableWidth());
-    style->set(Sid::pageEvenLeftMargin,   pf.evenLeftMargin());
-    style->set(Sid::pageOddLeftMargin,    pf.oddLeftMargin());
-    style->set(Sid::pageEvenTopMargin,    pf.evenTopMargin());
-    style->set(Sid::pageEvenBottomMargin, pf.evenBottomMargin());
-    style->set(Sid::pageOddTopMargin,     pf.oddTopMargin());
-    style->set(Sid::pageOddBottomMargin,  pf.oddBottomMargin());
-    style->set(Sid::pageTwosided,         pf.twosided());
-}
-
-//---------------------------------------------------------
-//   initPageFormat
-//    initialize PageFormat from Style
-//---------------------------------------------------------
-
-void initPageFormat(MStyle* style, PageFormat* pf)
-{
-    SizeF sz;
-    sz.setWidth(style->value(Sid::pageWidth).toReal());
-    sz.setHeight(style->value(Sid::pageHeight).toReal());
-    pf->setSize(sz);
-    pf->setPrintableWidth(style->value(Sid::pagePrintableWidth).toReal());
-    pf->setEvenLeftMargin(style->value(Sid::pageEvenLeftMargin).toReal());
-    pf->setOddLeftMargin(style->value(Sid::pageOddLeftMargin).toReal());
-    pf->setEvenTopMargin(style->value(Sid::pageEvenTopMargin).toReal());
-    pf->setEvenBottomMargin(style->value(Sid::pageEvenBottomMargin).toReal());
-    pf->setOddTopMargin(style->value(Sid::pageOddTopMargin).toReal());
-    pf->setOddBottomMargin(style->value(Sid::pageOddBottomMargin).toReal());
-    pf->setTwosided(style->value(Sid::pageTwosided).toBool());
-}
-
-//---------------------------------------------------------
-//   readPageFormat
-//---------------------------------------------------------
-
-void readPageFormat(MStyle* style, XmlReader& e)
-{
-    PageFormat pf;
-    initPageFormat(style, &pf);
-    pf.read(e);
-    setPageFormat(style, pf);
-}
 
 //---------------------------------------------------------
 //   readTextStyle206
@@ -3377,7 +3327,7 @@ static void readStyle(MStyle* style, XmlReader& e)
         } else if (tag == "Spatium") {
             style->set(Sid::spatium, e.readDouble() * DPMM);
         } else if (tag == "page-layout") {
-            readPageFormat(style, e);
+            compat::readPageFormat206(style, e);
         } else if (tag == "displayInConcertPitch") {
             style->set(Sid::concertPitch, QVariant(bool(e.readInt())));
         } else if (tag == "pedalY") {
@@ -3710,73 +3660,6 @@ static bool readScore(Score* score, XmlReader& e)
     }
 
     return true;
-}
-
-//---------------------------------------------------------
-//   read
-//  <page-layout>
-//      <page-height>
-//      <page-width>
-//      <landscape>1</landscape>
-//      <page-margins type="both">
-//         <left-margin>28.3465</left-margin>
-//         <right-margin>28.3465</right-margin>
-//         <top-margin>28.3465</top-margin>
-//         <bottom-margin>56.6929</bottom-margin>
-//         </page-margins>
-//      </page-layout>
-//---------------------------------------------------------
-
-void PageFormat::read(XmlReader& e)
-{
-    qreal _oddRightMargin  = 0.0;
-    qreal _evenRightMargin = 0.0;
-    QString type;
-
-    while (e.readNextStartElement()) {
-        const QStringRef& tag(e.name());
-        if (tag == "page-margins") {
-            type = e.attribute("type", "both");
-            qreal lm = 0.0, rm = 0.0, tm = 0.0, bm = 0.0;
-            while (e.readNextStartElement()) {
-                const QStringRef& t(e.name());
-                qreal val = e.readDouble() * 0.5 / PPI;
-                if (t == "left-margin") {
-                    lm = val;
-                } else if (t == "right-margin") {
-                    rm = val;
-                } else if (t == "top-margin") {
-                    tm = val;
-                } else if (t == "bottom-margin") {
-                    bm = val;
-                } else {
-                    e.unknown();
-                }
-            }
-            _twosided = type == "odd" || type == "even";
-            if (type == "odd" || type == "both") {
-                _oddLeftMargin   = lm;
-                _oddRightMargin  = rm;
-                _oddTopMargin    = tm;
-                _oddBottomMargin = bm;
-            }
-            if (type == "even" || type == "both") {
-                _evenLeftMargin   = lm;
-                _evenRightMargin  = rm;
-                _evenTopMargin    = tm;
-                _evenBottomMargin = bm;
-            }
-        } else if (tag == "page-height") {
-            _size.setHeight(e.readDouble() * 0.5 / PPI);
-        } else if (tag == "page-width") {
-            _size.setWidth(e.readDouble() * .5 / PPI);
-        } else {
-            e.unknown();
-        }
-    }
-    qreal w1        = _size.width() - _oddLeftMargin - _oddRightMargin;
-    qreal w2        = _size.width() - _evenLeftMargin - _evenRightMargin;
-    _printableWidth = qMin(w1, w2);       // silently adjust right margins
 }
 
 //---------------------------------------------------------
