+++ src/engraving/libmscore/read206.h
@@ -28,59 +28,8 @@
 namespace Ms {
 class MStyle;
 
-//---------------------------------------------------------
-//   @@ PageFormat
-//---------------------------------------------------------
-
-class PageFormat
-{
-    mu::SizeF _size;
-    qreal _printableWidth;          // _width - left margin - right margin
-    qreal _evenLeftMargin;          // values in inch
-    qreal _oddLeftMargin;
-    qreal _evenTopMargin;
-    qreal _evenBottomMargin;
-    qreal _oddTopMargin;
-    qreal _oddBottomMargin;
-    bool _twosided;
-
-public:
-    PageFormat() {}
-
-    const mu::SizeF& size() const { return _size; }                        // size in inch
-    qreal width() const { return _size.width(); }
-    qreal height() const { return _size.height(); }
-    void setSize(const mu::SizeF& s) { _size = s; }
-
-    void read(XmlReader&);
-    qreal evenLeftMargin() const { return _evenLeftMargin; }
-    qreal oddLeftMargin() const { return _oddLeftMargin; }
-    qreal evenTopMargin() const { return _evenTopMargin; }
-    qreal evenBottomMargin() const { return _evenBottomMargin; }
-    qreal oddTopMargin() const { return _oddTopMargin; }
-    qreal oddBottomMargin() const { return _oddBottomMargin; }
-    qreal printableWidth() const { return _printableWidth; }
-
-    void setEvenLeftMargin(qreal val) { _evenLeftMargin = val; }
-    void setOddLeftMargin(qreal val) { _oddLeftMargin = val; }
-    void setEvenTopMargin(qreal val) { _evenTopMargin = val; }
-    void setEvenBottomMargin(qreal val) { _evenBottomMargin = val; }
-    void setOddTopMargin(qreal val) { _oddTopMargin = val; }
-    void setOddBottomMargin(qreal val) { _oddBottomMargin = val; }
-    void setPrintableWidth(qreal val) { _printableWidth = val; }
-
-    bool twosided() const { return _twosided; }
-    void setTwosided(bool val) { _twosided = val; }
-
-    // convenience functions
-    qreal evenRightMargin() const { return _size.width() - _printableWidth - _evenLeftMargin; }
-    qreal oddRightMargin() const { return _size.width() - _printableWidth - _oddLeftMargin; }
-};
-
 extern Element* readArticulation(Element*, XmlReader&);
 extern void readAccidental206(Accidental*, XmlReader&);
-extern void setPageFormat(MStyle*, const PageFormat&);
-extern void initPageFormat(MStyle*, PageFormat*);
 extern void readTextStyle206(MStyle* style, XmlReader& e, std::map<QString, std::map<Sid, QVariant> >& excessStyles);
 //extern void readText206(XmlReader& e, TextBase* t, Element* be);
 // extern void readVolta206(XmlReader& e, Volta* volta);
